floating point added , but postgresql not work float with fraction

+++ migrations/20220329081238-account_balances.js
@@ -17,7 +17,7 @@ module.exports = {
         allowNull: true
       },      
       balance_amount:{
-        type:Sequelize.BIGINT,
+        type:Sequelize.FLOAT(10, 2),
         defaultValue:0,
         allowNull:false
       },
